feat(ios): add "Load Sample Data" settings option for evaluators
Adds a Settings affordance (test builds only) that populates the app with
a realistic ~90-day dataset so evaluators can explore the app without
hand-entering data.

- Extract the existing screenshot dataset into a reusable `SampleDataLoader`,
  shared by the `--load-screenshot-data` launch arg and the new UI.
- Add `BuildEnvironment.isPreRelease` (DEBUG or TestFlight sandbox receipt)
  to gate the option out of App Store production builds.
- New `SampleDataSectionView`: backs up current data first (aborts if the
  backup fails), then resets and loads the sample dataset, and points users
  to Backup & Export to restore.
- Add SampleDataLoader unit tests.
